Moved centralized error messages to a new module
